make-clean.output: Cleaned up the remaining drive by hand again.
It looks like instead of \.\CDROM12 it should have been \.\CDROM4 to be unmounted.  I did that by hand and repeated the same command several times in a row until it started throwing error:

    Dismount-DiskImage : The system cannot find the file specified.

That was pretty unexpected.  Probably has to do with me playing with `RemoveDrive_x64\RemoveDrive.exe M:` several times in row to restore the needed set of drive letters assigned to the mounted virtual drives.